++ User Password Encryption by seed.

+++ src/auto/controller/meta.module.controller.js
@@ -17,10 +17,11 @@
 
 import asyncHandler from "../../shared/middleware/asyncHandler";
 import {MetaModule, ModuleStatus} from '../models'
-import schemaDesignFromMetaModule from "../../shared/helpers/schema.helper";
+import schemaDesignFromMetaModule, {getDynamicModuleByUrl} from "../../shared/helpers/schema.helper";
 import mongoose from "mongoose";
 import {fullCamelCase} from "../../shared/helpers/util.fun";
 import pluralize from 'pluralize';
+import {EN_ModuleStatus} from "../models/app-data";
 
 // @desc  Get All schemas
 // @method GET
@@ -41,7 +42,7 @@ export const CreateMetaModules = asyncHandler(async (req, res, next) => {
         moduleName = fullCamelCase(moduleName);
         const moduleId = pluralize(moduleName).toLowerCase();
         moduleName = pluralize.singular(moduleName);
-        const newObject = {clientId, projectId, status: 'DRAFT', moduleId, moduleName, fields};
+        const newObject = {clientId, projectId, status: EN_ModuleStatus.DRAFT, moduleId, moduleName, fields};
         const addStatus = await MetaModule.create(newObject);
         res.status(201).json({success: true, data: addStatus})
     } else {
@@ -64,21 +65,17 @@ export const ClearAllMetaModules = asyncHandler(async (req, res, next) => {
 // @method PUT
 // @url /metamodules/publish/:id
 export const PublishTheModule = asyncHandler(async (req, res, next) => {
-    if (req.body) {
         const id = req.params.id;
-        const status = req.body.status;
         const foundDataModel = await MetaModule.findById(id);
         const sch =  schemaDesignFromMetaModule(foundDataModel);
         const checkedThenGetDynamicModule = await mongoose.models[foundDataModel.moduleName] || await mongoose.model(foundDataModel.moduleName, sch);
-        let updateStatus =  false;
         if(checkedThenGetDynamicModule){
-             updateStatus = await foundDataModel.updateOne({_id: id}, { $set: { status: status } });
+            const updateStatus = await MetaModule.updateOne({_id: id}, {status: EN_ModuleStatus.PUBLISHED});
+            res.status(201).json({success: true, data: updateStatus});
+        }else {
+            next(Error('Please verify the Id.!'));
         }
-        // Update all documents in the `mymodels` collection
-        res.status(201).json({success: true, data: updateStatus})
-    } else {
-        next(Error('Please provide data..!'));
-    }
+
 });
 
 
@@ -100,7 +97,7 @@ export const GetMetaModulesByIdHandler = asyncHandler(async (req, res, next) => 
 // @url /metamodules/:id
 export const UpdateMetaModulesByIdHandler = asyncHandler(async (req, res, next) => {
     const {clientId, moduleId, moduleName, fields} = req.body;
-    const newObject = {clientId, moduleId, moduleName, fields}
+    const newObject = {clientId, moduleId, moduleName, fields, status: EN_ModuleStatus.DRAFT}
     const id = req.params.id;
     const addStatus = await MetaModule.findOneAndReplace({_id: id}, {...newObject});
     res.status(201).json({success: true, data: addStatus})
@@ -109,9 +106,29 @@ export const UpdateMetaModulesByIdHandler = asyncHandler(async (req, res, next) 
 
 // // @desc  Delete Handler Function
 // // @method delete
-// // @url /metamodules/:id
+// // @url /metamodules/fdelete/:id
 export const DeleteMetaModulesByIdHandler = asyncHandler(async (req, res, next) => {
     const id = req.params.id;
-    const addStatus = await MetaModule.deleteOne({_id: id});
-    res.status(200).json({success: true, message: `Successfully deleted ${id}`})
+    // find record by id
+    const foundDataModel = await MetaModule.findById(id);
+    // remove respective collection
+    const dynamicModule = await getDynamicModuleByUrl(foundDataModel.moduleId);
+    const deleteStatus = await  dynamicModule.collection.drop();
+    // remove the entry from the main module collection
+    if(deleteStatus){
+        const addStatus = await MetaModule.deleteOne({_id: id});
+        res.status(200).json({success: true, message: `Successfully deleted ${id}`})
+    }else{
+        next(Error('There is some problem to delete the data!'));
+    }
+});
+
+
+// @desc Soft delete  Handler Function
+// @method DELETE
+// @url /metamodules/inactive/:id
+export const SoftDeleteMetaModulesByIdHandler = asyncHandler(async (req, res, next) => {
+    const id = req.params.id;
+    const updateStatus = await MetaModule.findOneAndReplace({_id: id}, {isActive: false});
+    res.status(201).json({success: true, data: updateStatus})
 });
